+++ quiz-app/documentation/change_log.md
@@ -5,3 +5,8 @@
 - New user can be created and it is stored to database
 - User can see user page after login
 - User can logout
+  
+## Week4
+- Added questions tables and created database connection
+- Added QuestionsRepository class, which handles saving the new questions to SQLite database
+